chore: move related type methods to types.js
